+++ src/program/Signature.cpp
@@ -296,8 +296,14 @@ int SigSearch::SearchAVX2(uint8_t* input, size_t inputLen, const Signature &sig,
 // Search for signature pattern, returning a status result
 int SigSearch::Search(uint8_t* input, size_t inputLen, const Signature &sig, ptrdiff_t* output_offset)
 {
+#ifdef __arch64__
+    /* AVX2 is for x86 arch, but a similar function could be written for aarch64 using SIMD:
+     * <http://0x80.pl/notesen/2016-11-28-simd-strfind.html#aarch64-64-bit-code>
+     */
+    static bool isAVX2Supported = false;
+#else
     static bool isAVX2Supported = __builtin_cpu_supports("avx2");
-    
+#endif
     if (isAVX2Supported)
         return SearchAVX2(input, inputLen, sig, output_offset);
     else